==> codes/nb2.2_diff_dataset_size.ipynb <==
import pandas as pd
dr = '~/PycharmProjects/code-trans-aug/storage/dataset_large/'
orig = pd.read_csv(dr+'dataset_org.csv')
aug = pd.read_csv(dr+'dataset_aug.csv')
orig.shape, aug.shape

sam = orig.sample(frac=0.33)
sam.to_csv(dr+'dataset_org_one_third.csv', header=None, index=False)

dataset = pd.concat([sam, aug.sample(int(len(sam)*0.25))])
dataset.to_csv(dr+'dataset_or_aug_one_third.csv', header=None, index=False)
