==> codes/nb3_split_dataset_c4.ipynb <==
dr = '~/PycharmProjects/code-trans-aug/storage/alternative/'

d_t = 'c4/'
cols = ['id1', 'id2', 'label']
import pandas as pd

def split_dataset(data):
    ratios = [8,1,1]

    data_num = len(data)
    train_split = int(ratios[0] / sum(ratios) * data_num)
    val_split = train_split + int(ratios[1] / sum(ratios) * data_num)

    data = data.sample(frac=1, random_state=666)
    
    train_list = data.iloc[:train_split]
    valid_list = data.iloc[train_split:val_split]
    test_list = data.iloc[val_split:]
    
    return train_list, valid_list, test_list


original_pairs = pd.read_csv(dr+'clones.csv', names=cols)
augmented_pairs = pd.read_csv(dr+'dataset_or_80_20_aug.csv', names=cols)
augmented_pairs = augmented_pairs[augmented_pairs.label==1]


train, dev, test = split_dataset(original_pairs)
train.to_csv(dr+d_t+'train.csv',header=None, index=False)
dev.to_csv(dr+d_t+'dev.csv',header=None, index=False)
test.to_csv(dr+d_t+'test.csv',header=None, index=False)

train_aug, dev_aug, test_aug = split_dataset(augmented_pairs)
train_aug.to_csv(dr+d_t+'train_aug.csv',header=None, index=False)
dev_aug.to_csv(dr+d_t+'dev_aug.csv',header=None, index=False)
test_aug.to_csv(dr+d_t+'test_aug.csv',header=None, index=False)

original_pairs.shape

augmented_pairs.shape
