==> codes/nb5_dataset_stats.ipynb <==
import pandas as pd
dr = '~/PycharmProjects/code-trans-aug/storage/'

def get_info(data):


    cj = 0
    cp = 0
    for i, r in data.iterrows():
        if r.language_extension=='java':
            cj+=len(r.code.split(';'))
        else:
            cp+=len(r.code.split('\n'))
    #print(cj, cp)

    return  data.problem_class.nunique(), cj/len(data), cp/len(data), len(data[data.is_parsable==1]), len(data[data.is_parsable==0])

data = pd.read_pickle(dr+'data_transpiled.pkl')

p = data[data.language_extension=='py']
j = data[data.language_extension=='java']


for d in [j,p]:
    at = pd.DataFrame([r for i, r in d.iterrows() if 'at_coder' in r.file_path])
    gcj = pd.DataFrame([r for i, r in d.iterrows() if 'code_jam' in r.file_path])
    for df in [at, gcj]:
        print(get_info(df))



d = '~/PycharmProjects/code-trans-aug/storage/dataset_large/'
ol = pd.read_csv(d+'dataset_org.csv')
ola = pd.read_csv(d+'dataset_aug.csv')
ol82 = pd.read_csv(d+'dataset_or_80_20_aug.csv')

s = '~/PycharmProjects/code-trans-aug/storage/dataset_small/'
os = pd.read_csv(s+'dataset_org.csv')
osa = pd.read_csv(s+'dataset_aug.csv')
os82 = pd.read_csv(s+'dataset_or_80_20_aug.csv')

ol.shape, ola.shape, ol82.shape,'===', os.shape, osa.shape , os82.shape


data = pd.read_pickle(dr+'alternative/codes_alt.pkl')

data = data.drop(['is_parsable'], axis=1)

data

def is_parsable(text, ext):
    f = False
    if text != -1:
        if ext=='py':
            try:
                _ = ast.parse(text)
                f = True
                #dataset.append([fp, pc, text, ext, 1])
            except:
                f = False
                #dataset.append([fp, pc, text, ext, 0])
        elif ext=='java':
            try:
                programtokens = javalang.tokenizer.tokenize(text)
                parser = javalang.parse.Parser(programtokens)
                _ = parser.parse_member_declaration()
                f = True
            except:
                try:
                    programtokens = javalang.tokenizer.tokenize(text)
                    _ = javalang.parser.parse(programtokens)
                    f = True
                except:
                    f = False
        return f
x = []
for i, row in data.iterrows():
    f = is_parsable(row.code, row.language_extension)
    #print(f)
    if f:
        x.append(list(row)+[1])
    else:
        x.append(list(row)+[0])



cc = list(data.columns) + ['is_parsable']

alt_data_up = pd.DataFrame(x, columns=cc)

def get_info(data):
    

    cj = 0
    cp = 0
    for i, r in data.iterrows():
        if r.language_extension=='java':
            cj+=len(r.code.split(';'))
        else:
            cp+=len(r.code.split('\n'))
    #print(cj, cp)

    return  data.problem_class.nunique(), cj/len(data), cp/len(data), len(data[data.is_parsable==1]), len(data[data.is_parsable==0])


data = alt_data_up

p = data[data.language_extension=='py']
j = data[data.language_extension=='java']


for d in [j,p]:
    at = pd.DataFrame([r for i, r in d.iterrows() if 'at_coder' in r.file_path])
    gcj = pd.DataFrame([r for i, r in d.iterrows() if 'code_jam' in r.file_path])
    for df in [at, gcj]:
        print(get_info(df))
